pyproject.toml
[project]
name = "pubmed-paper-fetcher-cl"
version = "0.1.3"
description = "Fetch PubMed papers and filter for pharma/biotech authors"
authors = [
    { name = "ChaahatL", email = "chaahatlokhande@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "biopython (>=1.85,<2.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "pubmed_cli" },
    { include = "scripts" }
]

[tool.poetry.scripts]
get-papers-list = "scripts.cli:main"

[tool.poetry.group.dev.dependencies]
twine = "^6.1.0"
pytest = "^8.4.1"
